+++ backend/src/knowledge_table_api/api/v1/endpoints/graph.py
@@ -11,10 +11,14 @@ from fastapi.encoders import jsonable_encoder
 from pydantic import ValidationError
 from whyhow import Chunk, ChunkMetadata
 
+from backend.src.knowledge_table_api.services.graph_service import (
+    generate_triples,
+)
+from backend.src.knowledge_table_api.services.llm_service import (
+    generate_schema,
+)
 from knowledge_table_api.core.dependencies import get_llm_service
-from knowledge_table_api.models.graph import Table
-from knowledge_table_api.services.graph import generate_triples
-from knowledge_table_api.services.llm_operations import generate_schema
+from knowledge_table_api.routing_schemas.graph import Table
 
 router = APIRouter(tags=["Graph"], prefix="/graph")