+++ backend/src/knowledge_table_api/models/query.py
@@ -1,6 +1,6 @@
 """Query model."""
 
-from typing import List, Literal, Optional, Union
+from typing import List, Literal, Optional
 
 from pydantic import BaseModel
 
@@ -11,27 +11,6 @@ class Rule(BaseModel):
     type: Literal["must_return", "may_return", "max_length"]
     options: Optional[List[str]] = None
     length: Optional[int] = None
-
-
-class QueryPrompt(BaseModel):
-    """Query prompt model."""
-
-    id: str
-    query: str
-    type: Literal["int", "str", "bool", "int_array", "str_array"]
-    entity_type: str
-    rules: Optional[List[Rule]] = None
-
-
-class QueryRequest(BaseModel):
-    """Query request model."""
-
-    document_id: str
-    previous_answer: Optional[Union[int, str, bool, List[int], List[str]]] = (
-        None
-    )
-    prompt: QueryPrompt
-    rag_type: Optional[Literal["vector", "hybrid", "decomposed"]] = "hybrid"
 
 
 class Chunk(BaseModel):
@@ -50,10 +29,3 @@ class Answer(BaseModel):
     answer: int | str | bool | List[int] | List[str] | None
     chunks: List[Chunk]
     type: Literal["int", "str", "bool", "int_array", "str_array"]
-
-
-class VectorResponse(BaseModel):
-    """Vector response model."""
-
-    message: str
-    chunks: List[Chunk]